Changes to ParNetwork2BBS and ParallelNetManager coming along.  Power outage may have lost some data.  next step link with SpikeSim

+++ BBS.h
@@ -118,7 +118,7 @@ public:
 	double threshold();
 	void cell();
 	void outputcell(int);
-	void spike_record(int, std::vector<float>,std::vector<float>);
+	void spike_record(int, std::vector<double>,std::vector<double>);
 	void netpar_solve(double);
 	void** gid2obj(int);
 	void** gid2cell(int);
